Implemented Maven, built jar file in local repository and added plugins.
