+++ src/main/java/com/solvd/zoo/animal/Mammal.java
@@ -1,0 +1,14 @@
+package com.solvd.zoo.animal;
+
+public abstract class Mammal extends Animal {
+
+    private int numberLegs;
+
+    public int getNumberLegs() {
+        return numberLegs;
+    }
+
+    public void setNumberLegs(int numberLegs) {
+        this.numberLegs = numberLegs;
+    }
+}
