+++ src/main/java/com/solvd/zoo/enums/Sex.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.enums;
+
+public enum Sex {
+
+    MALE,
+    FEMALE
+    
+}
