+++ src/main/java/com/solvd/zoo/exceptions/ExpiredTicketException.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.exceptions;
+
+public class ExpiredTicketException extends Exception {
+
+    public ExpiredTicketException(String message) {
+        super(message);
+    }
+}
