+++ src/main/java/com/solvd/zoo/exceptions/FeedAnimalException.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.exceptions;
+
+public class FeedAnimalException extends Exception {
+
+    public FeedAnimalException(String message) {
+        super(message);
+    }
+}
