+++ src/main/java/com/solvd/zoo/exceptions/InvalidNameException.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.exceptions;
+
+public class InvalidNameException extends Exception {
+
+    public InvalidNameException(String message) {
+        super(message);
+    }
+}
