+++ src/main/java/com/solvd/zoo/exceptions/LocationException.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.exceptions;
+
+public class LocationException extends Exception {
+
+    public LocationException(String message) {
+        super(message);
+    }
+}
