+++ src/main/java/com/solvd/zoo/exceptions/SpeakingException.java
@@ -1,0 +1,8 @@
+package com.solvd.zoo.exceptions;
+
+public class SpeakingException extends Exception {
+
+    public SpeakingException(String message) {
+        super(message);
+    }
+}
