+++ src/main/java/com/solvd/zoo/interfaces/IAttack.java
@@ -1,0 +1,5 @@
+package com.solvd.zoo.interfaces;
+
+public interface IAttack {
+    void attack();
+}
