+++ src/main/java/com/solvd/zoo/interfaces/IComplain.java
@@ -1,0 +1,10 @@
+package com.solvd.zoo.interfaces;
+
+import com.solvd.zoo.Zoo;
+import com.solvd.zoo.exceptions.SpeakingException;
+
+public interface IComplain {
+
+    void complain(String complainString, Zoo zoo) throws SpeakingException;
+
+}
