+++ src/main/java/com/solvd/zoo/interfaces/IDance.java
@@ -1,0 +1,5 @@
+package com.solvd.zoo.interfaces;
+
+public interface IDance {
+    void dance();
+}
