+++ src/main/java/com/solvd/zoo/interfaces/IPlay.java
@@ -1,0 +1,5 @@
+package com.solvd.zoo.interfaces;
+
+public interface IPlay {
+    void play();
+}
