+++ src/main/java/com/solvd/zoo/interfaces/IShout.java
@@ -1,0 +1,9 @@
+package com.solvd.zoo.interfaces;
+
+import com.solvd.zoo.exceptions.SpeakingException;
+
+public interface IShout {
+
+    void shout(String shoutString) throws SpeakingException;
+
+}
