--- zoo/animal/Reptile.java
@@ -1,14 +1,0 @@
-package zoo.animal;
-
-public abstract class Reptile extends Animal {
-
-    private boolean isVenomous;
-
-    public boolean isVenomous() {
-        return isVenomous;
-    }
-
-    public void setIsVenomous(boolean isVenomous) {
-        this.isVenomous = isVenomous;
-    }
-}
