--- zoo/enums/Sex.java
@@ -1,8 +1,0 @@
-package zoo.enums;
-
-public enum Sex {
-
-    MALE,
-    FEMALE
-    
-}
