--- zoo/exceptions/ExpiredTicketException.java
@@ -1,8 +1,0 @@
-package zoo.exceptions;
-
-public class ExpiredTicketException extends Exception {
-
-    public ExpiredTicketException(String message) {
-        super(message);
-    }
-}
