--- zoo/exceptions/FeedAnimalException.java
@@ -1,8 +1,0 @@
-package zoo.exceptions;
-
-public class FeedAnimalException extends Exception {
-
-    public FeedAnimalException(String message) {
-        super(message);
-    }
-}
