--- zoo/exceptions/InvalidNameException.java
@@ -1,8 +1,0 @@
-package zoo.exceptions;
-
-public class InvalidNameException extends Exception {
-
-    public InvalidNameException(String message) {
-        super(message);
-    }
-}
