--- zoo/exceptions/LocationException.java
@@ -1,8 +1,0 @@
-package zoo.exceptions;
-
-public class LocationException extends Exception {
-
-    public LocationException(String message) {
-        super(message);
-    }
-}
