--- zoo/exceptions/SpeakingException.java
@@ -1,8 +1,0 @@
-package zoo.exceptions;
-
-public class SpeakingException extends Exception {
-
-    public SpeakingException(String message) {
-        super(message);
-    }
-}
