--- zoo/interfaces/IAttack.java
@@ -1,5 +1,0 @@
-package zoo.interfaces;
-
-public interface IAttack {
-    void attack();
-}
