--- zoo/interfaces/IComplain.java
@@ -1,10 +1,0 @@
-package zoo.interfaces;
-
-import zoo.Zoo;
-import zoo.exceptions.SpeakingException;
-
-public interface IComplain {
-
-    void complain(String complainString, Zoo zoo) throws SpeakingException;
-
-}
