--- zoo/interfaces/IDance.java
@@ -1,5 +1,0 @@
-package zoo.interfaces;
-
-public interface IDance {
-    void dance();
-}
