--- zoo/interfaces/IPlay.java
@@ -1,5 +1,0 @@
-package zoo.interfaces;
-
-public interface IPlay {
-    void play();
-}
