--- zoo/interfaces/IShout.java
@@ -1,9 +1,0 @@
-package zoo.interfaces;
-
-import zoo.exceptions.SpeakingException;
-
-public interface IShout {
-
-    void shout(String shoutString) throws SpeakingException;
-
-}
