--- zoo/linkedlist/CustomNode.java
@@ -1,20 +1,0 @@
-package zoo.linkedlist;
-
-public class CustomNode<T> {
-
-    public T data;
-    public CustomNode<T> next;
-
-    public CustomNode(T data, CustomNode<T> next) {
-        this.data = data;
-        this.next = next;
-    }
-
-    public T getData() {
-        return this.data;
-    }
-
-    public void setData(T data) {
-        this.data = data;
-    }
-}
